requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-Migrate==1.2.0
Flask-SQLAlchemy==1.0
Flask-Script==2.0.5
Flask-Testing==0.4.2
Flask-WTF==0.10.2
base58check
pyjwt
ecdsa
requests
